requirements.txt
torch>=2.0.0
diffusers>=0.19.3
transformers>=4.31.0
gradio>=3.50.2
pillow>=9.5.0
numpy>=1.24.0 